+++ lib/default-es5-plugins.js
@@ -1,0 +1,15 @@
+"use strict";
+
+module.exports = [
+  ['transform-es2015-template-literals', {loose: true}],
+  ['transform-es2015-arrow-functions'],
+  ['transform-es2015-destructuring', {loose: true}],
+  ['transform-es2015-spread', {loose: true}],
+  ['transform-es2015-parameters'],
+  ['transform-es2015-computed-properties', {loose: true}],
+  ['transform-es2015-shorthand-properties'],
+  ['transform-es2015-block-scoping'],
+  ['check-es2015-constants'],
+  ['transform-es2015-classes', {loose: true}],
+  ['transform-proto-to-assign']
+];
