+++ lib/generate-external-helpers.js
