+++ test-support/index.html
@@ -11,7 +11,6 @@
   <script src="qunit.js"></script>
   <script src="/testem.js"></script>
   <script src="loader.js"></script>
-  <script src="babel-helpers.js"></script>
   <script src="vendor.js"></script>
   <script src="tests.js"></script>
   <script src="test-loader.js"></script>
